// src/model/policy-schema/policy.definitions.ts
export interface Evasion {
  /**
   * Human-readable name of sub-violation.
   */
  description: EvasionDescription;
  /**
   * Defines if sub-violation is enforced - alarmed or blocked, according to the 'Evasion
   * technique detected' (VIOL_EVASION) violation blocking settings.
   */
  enabled?: boolean;
  /**
   * Defines if sub-violation is learned. Sub-violations are learned only when learn is
   * enabled for the 'Evasion technique detected' (VIOL_EVASION) violation.
   */
  learn?: boolean;
  /**
   * Defines how many times the system decodes URI and parameter values before the request is
   * considered an evasion.
   * Relevant only for the 'Multiple decoding' sub-violation.
   */
  maxDecodingPasses?: number;
}

/**
 * Human-readable name of sub-violation.
 */
export enum EvasionDescription {
  ApacheWhitespace = "Apache whitespace",
  BadUnescape = "Bad unescape",
  BareByteDecoding = "Bare byte decoding",
  DirectoryTraversals = "Directory traversals",
  IISBackslashes = "IIS backslashes",
  IISUnicodeCodepoints = "IIS Unicode codepoints",
  MultipleDecoding = "Multiple decoding",
  UDecoding = "%u decoding",
}
// ...
export interface HTTPProtocol {
  /**
   * Human-readable name of sub-violation
   */
  description: HTTPProtocolDescription;
  /**
   * Defines if sub-violation is enforced - alarmed or blocked, according to the 'HTTP
   * protocol compliance failed' (VIOL_HTTP_PROTOCOL) violation blocking settings
   */
  enabled?: boolean;
  /**
   * Defines if sub-violation is learned. Sub-violations is learned only when learn is enabled
   * for the 'HTTP protocol compliance failed' (VIOL_HTTP_PROTOCOL) violation
   */
  learn?: boolean;
  /**
   * Defines maximum allowed number of headers in request.
   * Relevant only for the 'Check maximum number of headers' sub-violation
   */
  maxHeaders?: number;
  /**
   * Defines maximum allowed number of paramters in request.
   * Relevant only for the 'Check maximum number of parameters' sub-violation
   */
  maxParams?: number;
}

/**
 * Human-readable name of sub-violation
 */
export enum HTTPProtocolDescription {
  BadHTTPVersion = "Bad HTTP version",
  BadHostHeaderValue = "Bad host header value",
  BadMultipartFormDataRequestParsing = "Bad multipart/form-data request parsing",
  BadMultipartParametersParsing = "Bad multipart parameters parsing",
  BodyInGETOrHEADRequests = "Body in GET or HEAD requests",
  CRLFCharactersBeforeRequestStart = "CRLF characters before request start",
  CheckMaximumNumberOfHeaders = "Check maximum number of headers",
  CheckMaximumNumberOfParameters = "Check maximum number of parameters",
  ChunkedRequestWithContentLengthHeader = "Chunked request with Content-Length header",
  ContentLengthShouldBeAPositiveNumber = "Content length should be a positive number",
  HeaderNameWithNoHeaderValue = "Header name with no header value",
  HighASCIICharactersInHeaders = "High ASCII characters in headers",
  HostHeaderContainsIPAddress = "Host header contains IP address",
  MultipleHostHeaders = "Multiple host headers",
  NoHostHeaderInHTTP11Request = "No Host header in HTTP/1.1 request",
  NullInRequest = "Null in request",
  POSTRequestWithContentLength0 = "POST request with Content-Length: 0",
  SeveralContentLengthHeaders = "Several Content-Length headers",
  UnescapedSpaceInURL = "Unescaped space in URL",
  UnparsableRequestContent = "Unparsable request content",
}

export interface BlockingSettingsViolation {
  alarm?: boolean;
  block?: boolean;
  description?: string;
  learn?: boolean;
  name?: string;
}
// ...
export interface Class {
  action?: ClassAction;
  name: ClassName;
}

export enum ClassAction {
  Alarm = "alarm",
  Block = "block",
  Detect = "detect",
  Ignore = "ignore",
}

export enum ClassName {
  Malicious = "malicious",
  Trusted = "trusted",
  Untrusted = "untrusted",
}

/**
 * This section defines the properties of signature on the policy.
 */
export interface MitigationsSignature {
  action?: ClassAction;
  /**
   * The signature name, which along with the signature tag, identifies the signature.
   */
  name?: string;
}

export interface Settings {
  isEnabled?: boolean;
}
// ...
export enum HostNameTypeEnum {
  Explicit = "explicit",
  Wildcard = "wildcard",
}
// ...
export interface CookieSettings {
  /**
   * Maximum Cookie Header Length must be greater than 0 and less than 65536 bytes (64K).
   * Note: if 0 or *any* are set, then no restriction on the cookie header length is applied.
   */
  maximumCookieHeaderLength?: MaximumCookieHeaderLengthEnum | number;
}

export enum MaximumCookieHeaderLengthEnum {
  Any = "any",
}
// ...
export enum EnforcementType {
  Allow = "allow",
  Enforce = "enforce",
}
// ...
export enum InsertSameSiteAttribute {
  Lax = "lax",
  None = "none",
  NoneValue = "none-value",
  Strict = "strict",
}
// ...
export interface CSRFProtection {
  enabled?: boolean;
  expirationTimeInSeconds?: ExpirationTime | number;
  sslOnly?: boolean;
}

export enum ExpirationTime {
  Disabled = "disabled",
}

export interface CSRFURL {
  enforcementAction?: EnforcementAction;
  method?: CSRFURLMethod;
  parametersList?: string[];
  requiredParameters?: RequiredParameters;
  url: string;
  wildcardOrder?: number;
}

export enum EnforcementAction {
  None = "none",
  VerifyCSRFToken = "verify-csrf-token",
  VerifyOrigin = "verify-origin",
}

export enum CSRFURLMethod {
  Any = "any",
  Get = "GET",
  Post = "POST",
}

export enum RequiredParameters {
  AtLeastOne = "at-least-one",
  FromParametersList = "from-parameters-list",
  Ignore = "ignore",
}

/**
 * Data Guard feature can prevent responses from exposing sensitive information by masking
 * the data.
 */
export interface DataGuard {
  /**
   * If *true* the system considers credit card numbers as sensitive data.
   */
  creditCardNumbers?: boolean;
  /**
   * If *true* the system recognizes customized patterns as sensitive data.
   */
  customPatterns?: boolean;
  /**
   * List of PCRE regular expressions that specify the sensitive data patterns.
   */
  customPatternsList?: string[];
  /**
   * If *true* the system protects sensitive data.
   */
  enabled?: boolean;
  /**
   * Specifies the URLs for which the system enforces data guard protection.
   *
   * - **ignore-urls-in-list**: Specifies that the system enforces data guard protection for
   * all URLs except for those URLs in the Enforcement Mode list.
   * - **enforce-urls-in-list**: Specifies that the system enforces data guard protection only
   * for those URLs in the Enforcement Mode list
   */
  enforcementMode?: DataGuardEnforcementMode;
  /**
   * List of URLS to be enforced based on enforcement mode of data guard protection.
   */
  enforcementUrls?: string[];
  /**
   * If *true* the system recognizes exception patterns as not being sensitive data.
   */
  exceptionPatterns?: boolean;
  /**
   * List of PCRE regular expressions that specify the data patterns that are not recognized
   * as sensitive data.
   */
  exceptionPatternsList?: string[];
  /**
   * If *true* the system checks responses for file content.
   */
  fileContentDetection?: boolean;
  /**
   * File content types that the system considers to be sensitive data. Available file content
   * types:
   * - MS Office 2007 or later
   * - MS Office 97-2003 document
   * - PDF - Adobe Portable Document Format
   * - Mach-O - Mach object file format - used in Mac OS X
   * - PE - Portable Executable for Windows
   * - ELF - binary file format for Unix
   */
  fileContentDetectionFormats?: FileContentDetectionFormat[];
  /**
   * Specifies the number of last digits in credit card numbers that are exposed.
   */
  lastCcnDigitsToExpose?: number;
  /**
   * Specifies the number of last digits in U.S Social Security numbers that are exposed.
   */
  lastSsnDigitsToExpose?: number;
  /**
   * If *true* the system intercepts the returned responses to mask sensitive data.
   */
  maskData?: boolean;
  /**
   * If *true* the system considers U.S Social Security numbers as sensitive data.
   */
  usSocialSecurityNumbers?: boolean;
}

/**
 * Specifies the URLs for which the system enforces data guard protection.
 *
 * - **ignore-urls-in-list**: Specifies that the system enforces data guard protection for
 * all URLs except for those URLs in the Enforcement Mode list.
 * - **enforce-urls-in-list**: Specifies that the system enforces data guard protection only
 * for those URLs in the Enforcement Mode list
 */
export enum DataGuardEnforcementMode {
  EnforceUrlsInList = "enforce-urls-in-list",
  IgnoreUrlsInList = "ignore-urls-in-list",
}

export interface FileContentDetectionFormat {
  name?: FileContentDetectionFormatName;
}

export enum FileContentDetectionFormatName {
  ELFBinaryFileFormatForUnix = "ELF - binary file format for Unix",
  MSOffice2007OrLater = "MS Office 2007 or later",
  MSOffice972003Document = "MS Office 97-2003 document",
  MachOMachObjectFileFormatUsedInMACOSX = "Mach-O - Mach object file format - used in Mac OS X",
  PDFAdobePortableDocumentFormat = "PDF - Adobe Portable Document Format",
  PEPortableExecutableForWindows = "PE - Portable Executable for Windows",
}
// ...
export interface Header {
  allowRepeatedOccurrences?: boolean;
  base64Decoding?: boolean;
  checkSignatures?: boolean;
  htmlNormalization?: boolean;
  mandatory?: boolean;
  maskValueInLogs?: boolean;
  name: string;
  normalizationViolations?: boolean;
  percentDecoding?: boolean;
  signatureOverrides?: HeaderSignatureOverride[];
  type?: HostNameTypeEnum;
  urlNormalization?: boolean;
  wildcardOrder?: number;
}

export interface HeaderSignatureOverride {
  enabled?: boolean;
  name?: string;
  signatureId?: number;
  tag?: string;
}

export interface HostName {
  includeSubdomains?: boolean;
  name: string;
}
// ...
export interface MethodElement {
  actAsMethod?: ActAsMethod;
  name: string;
}

export enum ActAsMethod {
  Get = "GET",
  Post = "POST",
}
// ...
export interface OpenAPIFile {
  link: string;
}
// ...
export interface ServerTechnology {
  /**
   * Specifies the name of the selected policy. For example, PHP will add attack signatures
   * that cover known PHP vulnerabilities.
   */
  serverTechnologyName?: string;
}
// ...
export interface SignatureSet {
  /**
   * If enabled - when a signature from this signature set is detected in a request - the
   * request is logged.
   */
  alarm?: boolean;
  /**
   * If enabled - when a signature from this signature set is detected in a request (and the
   * signature is not in staging and the policy is in blocking mode) - the request is blocked.
   */
  block?: boolean;
  /**
   * If enabled - when a signature from this signature set is detected in a request -the
   * policy builder creates a learning suggestion to disable it.
   */
  learn?: boolean;
  /**
   * Signature set name.
   */
  name?: string;
  /**
   * Defines signature set.
   */
  signatureSet?: SignatureSetObject;
}

/**
 * Defines signature set.
 */
export interface SignatureSetObject {
  /**
   * Specifies filter that defines signature set.
   */
  filter?: SignatureSetFilter;
  signatures?: SignatureSetSignature[];
  type?: SignatureSetType;
}
// ...
export interface SignatureSetSignature {
  name?: string;
  signatureId?: number;
  tag?: string;
}

export enum SignatureSetType {
  FilterBased = "filter-based",
  Manual = "manual",
}
// ...
export interface PolicySignature {
  alarm?: boolean;
  block?: boolean;
  /**
   * Specifies, if true, that the signature is enabled on the security policy. When false, the
   * signature is disable on the security policy.
   */
  enabled?: boolean;
  inPolicy?: InPolicy;
  isPriorRuleEnforced?: boolean;
  learn?: boolean;
  /**
   * The signature name, which along with the signature tag, identifies the signature.
   */
  name?: string;
  /**
   * Specifies, if true, that the signature is in staging.
   * The system does not enforce signatures in staging. Instead, the system records the
   * request information and keeps it for a period of time
   * (the Enforcement Readiness Period whose default time period is 7 days).
   * Specifies, when false, that the staging feature is not in use, and that the system
   * enforces the signatures' Learn/Alarm/Block settings immediately.
   * (Blocking is performed only if the security policy's enforcement mode is Blocking.)
   */
  performStaging?: boolean;
  /**
   * The signature ID which identifies the signature.
   */
  signatureId?: number;
  /**
   * The signature tag, which along with the signature name, identifies the signature.
   */
  tag?: string;
}

export enum InPolicy {
  False = "false",
  True = "true",
}
// ...
export enum BlockRequests {
  Always = "always",
  Never = "never",
  PolicyDefault = "policy-default",
}